refactor: LSP(상속구조로 Cell 변경)
